Clarified pro-quints RFC, added RFC notes to table.
I have edited the multibase proquint RFC to clarify how the full prefix works, and I have added an explicit reference to the RFC in the table (because the proquint data encoded by multibase is different from the data encoded by the original proquint spec).

I have also added an explicit reference to the RFC for base8 in the table, because base8 according to the multibase RFC is significantly different from the base8 provided by other standard implementations (e.g. Python).